moved panel code to an own sub module
