+++ src/ncurses.rs
@@ -20,6 +20,7 @@ use std::{ char, ptr };
 use std::ffi::{CString, CStr};
 use self::ll::{FILE_p};
 pub use self::constants::*;
+pub use self::panel::wrapper::*;
 
 #[cfg(target_arch = "x86_64")]
 pub type chtype = u64;
@@ -32,6 +33,7 @@ pub type NCURSES_ATTR_T = attr_t;
 
 pub mod ll;
 pub mod constants;
+pub mod panel;
 
 trait FromCStr {
     fn from_c_str(s: *const libc::c_char) -> Self;
@@ -66,7 +68,6 @@ pub enum CURSOR_VISIBILITY
 
 pub type WINDOW = self::ll::WINDOW;
 pub type SCREEN = self::ll::SCREEN;
-pub type PANEL = self::ll::PANEL;
 pub type mmaskt = self::ll::mmask_t;
 pub type MEVENT = self::ll::MEVENT;
 
@@ -1909,59 +1910,3 @@ pub fn wmouse_trafo(w: *mut WINDOW, y: &mut[i32], x: &mut[i32], to_screen: bool)
 
 pub fn mouse_trafo( y: &mut[i32], x: &mut[i32], to_screen: bool  ) -> i32
 { unsafe { ll::mouse_trafo(y.as_mut_ptr(), x.as_mut_ptr(), to_screen as libc::c_int ) } }
-
-/*
- * Added panel support
- */
-
-#[cfg(feature="panel")]
-pub fn panel_window(panel: PANEL) -> WINDOW
-{ unsafe { ll::panel_window(panel) } }
-
-#[cfg(feature="panel")]
-pub fn update_panels()
-{ unsafe { ll::update_panels(); } }
-
-#[cfg(feature="panel")]
-pub fn hide_panel(panel: PANEL) -> i32
-{ unsafe { ll::hide_panel(panel) } }
-
-#[cfg(feature="panel")]
-pub fn show_panel(panel: PANEL) -> i32
-{ unsafe { ll::show_panel(panel) } }
-
-#[cfg(feature="panel")]
-pub fn del_panel(panel: PANEL) -> i32
-{ unsafe { ll::del_panel(panel) } }
-
-#[cfg(feature="panel")]
-pub fn top_panel(panel: PANEL) -> i32
-{ unsafe { ll::del_panel(panel) } }
-
-#[cfg(feature="panel")]
-pub fn bottom_panel(panel: PANEL) -> i32
-{ unsafe { ll::bottom_panel(panel) } }
-
-#[cfg(feature="panel")]
-pub fn new_panel(window: WINDOW) -> PANEL
-{ unsafe { ll::new_panel(window) } }
-
-#[cfg(feature="panel")]
-pub fn panel_above(panel: PANEL) -> PANEL
-{ unsafe { ll::panel_above(panel) } }
-
-#[cfg(feature="panel")]
-pub fn panel_below(panel: PANEL) -> PANEL
-{ unsafe { ll::panel_below(panel) } }
-
-#[cfg(feature="panel")]
-pub fn move_panel(panel: PANEL, y: i32, x: i32) -> i32
-{ unsafe { ll::move_panel(panel, y, x) } }
-
-#[cfg(feature="panel")]
-pub fn replace_panel(panel: PANEL, window: WINDOW) -> i32
-{ unsafe { ll::replace_panel(panel, window) } }
-
-#[cfg(feature="panel")]
-pub fn panel_hidden(panel: PANEL) -> bool
-{ unsafe { ll::panel_hidden(panel) == TRUE } }
